Made header for record structure
