+++ homework4/yellow_pages/record.h
@@ -1,0 +1,10 @@
+#pragma once
+
+const int nameLength = 30;
+const int numberLength = 15;
+
+struct Record
+{
+	char name[nameLength];
+	char number[numberLength];
+};
